student side restructured
