+++ src/components/Student/My_profile/Profile/Profile.js
@@ -1,5 +1,5 @@
 import React from "react";
-import "../Student.scss";
+import "../../../Student.scss";
 import StudentUpdateProfile from "./StudentUpdateProfile/StudentUpdateProfile";
 
 export default function Profile(props) {
